+++ src/Models/InvoiceLine.php
@@ -12,8 +12,12 @@ class InvoiceLine extends Model
     public $incrementing = false;
 
     protected $fillable = [
-        'amount', 'tax', 'tax_percentage', 'invoice_id', 'description', 'invoiceable_id', 'invoiceable_type',
-        'name', 'discount', 'quantity', 'is_free', 'is_complimentary'
+        'amount', 'tax', 'tax_details', 'invoice_id', 'description', 'invoiceable_id',
+        'invoiceable_type', 'name', 'discount', 'quantity', 'is_free', 'is_complimentary'
+    ];
+
+    protected $casts = [
+        'tax_details' => 'array'
     ];
 
     /**