+++ tests/Feature/BillTest.php
@@ -55,36 +55,115 @@ class BillTest extends AbstractTestCase
     }
 
     /** @test */
-    public function canAddAmountExclTaxToBill()
+    public function canAddAmountExclTaxPercentageToBill()
     {
         $this->service->create($this->customer);
 
-        $this->service->addAmountExclTax($this->product, 100, 'Some description', 0.21);
-        $this->service->addAmountExclTax($this->product, 100, 'Some description', 0.21);
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Some description');
 
         $this->assertEquals('242', (string) $this->service->getBill()->total);
         $this->assertEquals('42', (string) $this->service->getBill()->tax);
     }
 
     /** @test */
-    public function canAddAmountInclTaxToBill()
+    public function canAddAmountInclTaxPercentageToBill()
     {
         $new_bill = $this->service->create($this->customer)->getBill();
 
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
 
         $this->assertEquals('242', (string) $new_bill->total);
         $this->assertEquals('42', (string) $new_bill->tax);
     }
 
     /** @test */
+    public function canAddAmountExclTaxAmountToBill()
+    {
+        $this->service->create($this->customer);
+
+        $this->service->addTaxFixed('VAT', 21)->addAmountExclTax($this->product, 100, 'Some description');
+        $this->service->addTaxFixed('VAT', 21)->addAmountExclTax($this->product, 100, 'Some description');
+
+        $this->assertEquals('242', (string) $this->service->getBill()->total);
+        $this->assertEquals('42', (string) $this->service->getBill()->tax);
+    }
+
+    /** @test */
+    public function canAddAmountInclTaxAmountToBill()
+    {
+        $new_bill = $this->service->create($this->customer)->getBill();
+
+        $this->service->addTaxFixed('VAT', 21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxFixed('VAT', 21)->addAmountInclTax($this->product, 121, 'Some description');
+
+        $this->assertEquals('242', (string) $new_bill->total);
+        $this->assertEquals('42', (string) $new_bill->tax);
+    }
+
+    /** @test */
+    public function canAddAmountExclTaxAmountAndTaxPercentageToBill()
+    {
+        $this->service->create($this->customer);
+
+        $this->service->addTaxFixed('VAT', 21)->addAmountExclTax($this->product, 100, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Some description');
+
+        $this->assertEquals('242', (string) $this->service->getBill()->total);
+        $this->assertEquals('42', (string) $this->service->getBill()->tax);
+    }
+
+    /** @test */
+    public function canAddAmountInclTaxAmountAndTaxPercentageToBill()
+    {
+        $new_bill = $this->service->create($this->customer)->getBill();
+
+        $this->service->addTaxFixed('VAT', 21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
+
+        $this->assertEquals('242', (string) $new_bill->total);
+        $this->assertEquals('42', (string) $new_bill->tax);
+    }
+
+    /** @test */
+    public function canAddAmountExclMultipleTaxToBill()
+    {
+        $new_bill = $this->service->create($this->customer);
+
+        $this->service
+            ->addTaxFixed('TAX1', 1)
+            ->addTaxPercentage('TAX2', 0.21)
+            ->addTaxFixed('TAX3', 30)
+            ->addAmountExclTax($this->product, 100, 'Some description');
+
+        $this->assertEquals('152', (string) $new_bill->getBill()->total);
+        $this->assertEquals('52', (string) $new_bill->getBill()->tax);
+    }
+
+    /** @test */
+    public function canAddAmountInclMultipleTaxToBill()
+    {
+        $new_bill = $this->service->create($this->customer)->getBill();
+
+        $this->service
+            ->addTaxFixed('TAX1', 1)
+            ->addTaxPercentage('TAX2', 0.21)
+            ->addTaxFixed('TAX3', 30)
+            ->addAmountInclTax($this->product, 152, 'Some description');
+
+        $this->assertEquals('152', (string) $new_bill->total);
+        $this->assertEquals('52', (string) $new_bill->tax);
+    }
+
+
+    /** @test */
     public function canHandleNegativeAmounts()
     {
         $new_bill = $this->service->create($this->customer)->getBill();
 
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
-        $this->service->addAmountInclTax($this->product, -121, 'Some negative amount description', 0.21);
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, -121, 'Some negative amount description');
 
         $this->assertEquals('0', (string) $new_bill->total);
         $this->assertEquals('0', (string) $new_bill->tax);
@@ -105,8 +184,8 @@ class BillTest extends AbstractTestCase
     {
         $this->service->create($this->customer);
 
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
         $view = $this->service->view();
         $rendered = $view->render(); // fails if view cannot be rendered
         $this->assertTrue(true);
@@ -116,8 +195,8 @@ class BillTest extends AbstractTestCase
     public function canGetBillPdf()
     {
         $this->service->create($this->customer);
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
         $pdf = $this->service->pdf();  // fails if pdf cannot be rendered
         $this->assertTrue(true);
     }
@@ -126,8 +205,8 @@ class BillTest extends AbstractTestCase
     public function canDownloadBillPdf()
     {
         $this->service->create($this->customer);
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
-        $this->service->addAmountInclTax($this->product, 121, 'Some description', 0.21);
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Some description');
         $download = $this->service->download(); // fails if pdf cannot be rendered
         $this->assertTrue(true);
     }
@@ -172,8 +251,8 @@ class BillTest extends AbstractTestCase
     {
         $new_bill = $this->service->create($this->customer)->getBill();
 
-        $this->service->setFree()->addAmountExclTax($this->product, 100, 'Free sale', 0.21);
-        $this->service->addAmountExclTax($this->product, 100, 'Some description', 0.21);
+        $this->service->setFree()->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Free sale');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Some description');
 
         $this->assertEquals('121', (string) $new_bill->total);
         $this->assertEquals('21', (string) $new_bill->tax);
@@ -185,8 +264,8 @@ class BillTest extends AbstractTestCase
     {
         $new_bill = $this->service->create($this->customer)->getBill();
 
-        $this->service->setComplimentary()->addAmountExclTax($this->product, 100, 'Complimentary sale', 0.21);
-        $this->service->addAmountExclTax($this->product, 100, 'Some description', 0.21);
+        $this->service->setComplimentary()->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Complimentary sale');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Some description');
 
         $this->assertEquals('121', (string) $new_bill->total);
         $this->assertEquals('21', (string) $new_bill->tax);
@@ -198,8 +277,8 @@ class BillTest extends AbstractTestCase
     {
         $new_bill = $this->service->create($this->customer)->getBill();
 
-        $this->service->setComplimentary()->addAmountExclTax($this->product, 100, 'Complimentary sale', 0.21);
-        $this->service->setFree()->addAmountInclTax($this->product, 121, 'Free sale', 0.21);
+        $this->service->setComplimentary()->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Complimentary sale');
+        $this->service->setFree()->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Free sale');
 
         $this->assertEquals('0', (string) $new_bill->total);
         $this->assertEquals('0', (string) $new_bill->tax);
@@ -211,9 +290,9 @@ class BillTest extends AbstractTestCase
     {
         $new_bill = $this->service->create($this->customer)->getBill();
 
-        $this->service->setComplimentary()->addAmountExclTax($this->product, 100, 'Complimentary sale', 0.21);
-        $this->service->setFree()->addAmountInclTax($this->product, 121, 'Free sale', 0.21);
-        $this->service->addAmountInclTax($this->product, 121, 'Regular sale', 0.21);
+        $this->service->setComplimentary()->addTaxPercentage('VAT', 0.21)->addAmountExclTax($this->product, 100, 'Complimentary sale');
+        $this->service->setFree()->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Free sale');
+        $this->service->addTaxPercentage('VAT', 0.21)->addAmountInclTax($this->product, 121, 'Regular sale');
 
         $this->assertEquals('121', (string) $new_bill->total);
         $this->assertEquals('21', (string) $new_bill->tax);